+++ src/main/java/com/car/foryou/service/UserService.java
@@ -7,8 +7,8 @@ import org.springframework.data.domain.Page;
 
 public interface UserService {
     Page<UserResponse> getAllUsers(UserFilterParam userFilterParam);
-    UserResponse getUserById(long id);
+    UserResponse getUserById(int id);
     UserResponse createUser(UserRequest request);
-    UserResponse updateUser(long id, UserRequest request);
-    UserResponse deleteUser(long id);
+    UserResponse updateUser(int id, UserRequest request);
+    UserResponse deleteUser(int id);
 }
